Update README.md for the NameWrapper
Update the methods for checking whether a name is Unregistered / Unwrapped. When an Emancipated / Locked name expires, the NameWrapper will return `address(0)`, but the Registry will still have the NameWrapper set as the owner.

Update some of the details for the Expiry. There is a nuance between when a "name expires", and when fuses are reset. Only Emancipated / Locked names expire. However, the expiry still has a practical and useful effect on Wrapped names, as it determines when any parent-controlled fuses get reset.